Fix test compile error: 222 doesn't fit in a signed Byte literal
CI caught it: byteArrayOf(10, 111, 222, 5) fails because 222 exceeds
Byte's signed range (-128..127) as a literal - needs an explicit
.toByte() conversion (wraps to -34, the correct byte representation
of 222 unsigned, same value getByAddress expects).

+++ app/src/test/java/com/owner/assistant/vpn/DnsPacketProcessorTest.kt
@@ -14,7 +14,7 @@ import java.net.InetAddress
  */
 class DnsPacketProcessorTest {
 
-    private val clientIp = InetAddress.getByAddress(byteArrayOf(10, 111, 222, 5))
+    private val clientIp = InetAddress.getByAddress(byteArrayOf(10, 111, 222.toByte(), 5))
     private val dnsServerIp = InetAddress.getByAddress(byteArrayOf(8, 8, 8, 8))
     private val clientPort = 54321
